refac(src):  Restructure folders Restructured the directories in Frontend/src to be better organized and easier to navigate (or at least that's the idea anyways).  Updated import statements.  May add some more subdirectories later on if these feel too bloated with tests in them.
