+++ Frontend/src/ImageView/BottomToolbar/AddTagButton.js
