+++ Frontend/src/ImageView/BottomToolbar/ExcessTagsChip.js
