+++ Frontend/src/ImageView/BottomToolbar/Tag.js
@@ -2,7 +2,7 @@ import React, { useContext } from 'react';
 import { ThemeProvider, alpha, createTheme } from '@mui/material/styles';
 import Chip from '@mui/material/Chip'
 
-import AppDataContext from '../SupportingModules/AppDataContext';
+import AppDataContext from '../../SupportingModules/AppDataContext';
 
 
 /**
