+++ Frontend/src/ImageView/TopToolbar/DeleteImageButton.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { IconButton } from "@mui/material";
 import DeleteIcon from "@mui/icons-material/Delete";
 
-import filterSocket from '../SupportingModules/FilterSocket';
+import filterSocket from '../../SupportingModules/FilterSocket';
 
 
 /**
