+++ Frontend/src/MainPage/TagDrawer/TagEditor.js
@@ -3,7 +3,7 @@ import { FormControlLabel, TextField } from '@mui/material';
 import { useTheme } from '@mui/material/styles';
 import useMediaQuery from '@mui/material/useMediaQuery';
 
-import filterSocket from '../SupportingModules/FilterSocket';
+import filterSocket from '../../SupportingModules/FilterSocket';
 
 
 export default function TagEditor(props) {
