+++ Frontend/src/MainPage/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 
-import ApiCall from './SupportingModules/ApiCall';
+import ApiCall from '../SupportingModules/ApiCall';
 import App, { AppChildren } from './App';
